DOAXVV: Backup IB + VB before shadow shaders and restore afterwards
We're a bit lazy and don't backup the buffers for every draw call
because the game seemed to replace them anyway for the next draw call,
but this turns out not to be the case for the shadow map shaders, which
may draw from the same VB+IB multiple times in a row without rebinding
them, so if we have replaced the mesh on a single draw call it may
affect subsequent draw calls we didn't intend to.

Add a backup and restore of the VB+IB on all the shadow shaders before
calling out to the mods.