Unity Template: Migrate to split d3dx.ini configuration
This moves out the Unity template d3dx.ini excerpts into a separate
unity.ini file, and changes the scripts to use unity_generated.ini for
all automatically generated sections.

This should result in a much cleaner d3dx.ini, as all that should need
to be added is now:

    [Include]
    include = ShaderFixes/unity.ini

The cleanup script will now attempt to clean the split configuration if
it exists, and will only touch the d3dx.ini if the split configuration
doesn't yet exist. That should allow the d3dx.ini to get one final clean
when migrating, but will not touch it ever again. As an added bonus,
since the lighting sections are in a separate section the cleanup script
will never delete them.