asmtool: Fix Unity template broken by 3DMigoto v1.3.1 / v1.2.72
The 3DMigoto v1.3.1 and v1.2.72 updates started stripping hidden
whitespace from the end of lines in assembly shaders to make ShaderRegex
patterns easier to write, but this also broke the Unity lighting fix, as
asmtool searched for TEXCOORDs in the shader, matching this hidden
whitespace at the end of lines.

Adjust the regex used to match these lines to make the whitespace
optional.