DOAXVV: Update d3dx.ini & SBS shader from template, set track_texture_updates=2
Brings in the performance monitor key bindings and SLI optimised version
of the SBS shader from 1.3.11.

This also disables hash contamination detection in hunting mode - this
game does suffer from hash contamination in the UI (so a translation mod
would require hash tracking to be enabled), but not the meshes or
skin/costume textures where the modding efforts are focussed. Since this
is understood, and other modders are using this in areas where it does
not matter, we can turn this option off to save a few fps in hunting
mode.