Auto-convergence: Disable auto-convergence when 3D Vision is inactive
This uses the new conditional logic support in 3DMigoto 1.3.11 to
disable the auto-convergence shader when 3D Vision is inactive or
unavailable, which should help improve the performance for non 3D-Vision
users, who may have less powerful machines.

This also makes setting the convergence depend on auto-convergence being
enabled - this should ensure that the Burst scene will use the override
in the preset and will not get an auto-convergence value that was still
in flight from a previous frame when the Burst scene was active.