DOAXVV: Implement help text custom shader
This custom shader will load a text file and display it on the screen.

It will display a short message from help_short.txt for ten seconds
after the game has been launched, and will display the full help.txt
when F1 is pressed.

The shader's ini file can be extended to add additional text displays,
and uses a parameter buffer to tell the shader where to position the
text, how to anchor it, the size of the border, what colour it is and so
on. This is not the easiest way to write this information, but it avoids
taking up any ini params and still allows these parameters to be set on
a per-text box basis.

It is based on my earlier text shaders, but the length limitation is
removed by using a compute shader to process the text first, breaking it
up into smaller chunks that the geometry shaders invocations can render.
The compute shader will also wrap the text to fit within the width of
the bounds rectangle, and the geometry shader can anchor the text to
either side or center of the bounds rectangle.

The geometry shader's invocation 0 is used specially for the purpose of
rendering the background rectangle - I had some concerns that this might
flicker if the order was not consistent (which would just need a second
shader pass to solve), but it seems to work just fine.

Ini Param usage is kept to a minimum - x5 and y5 are used to store the
window width/height used by the shader to properly size the text and to
detect resolution changes that require the formatting compute shader to
be re-run.