DOAXVV: Add keyboard shortcuts to disable costume mods
The costume mods can take a bit of performance, since the
checktextureoverride commands get called a lot during each frame. This
adds a key binding to toggle them on and off to regain performance when
they aren't needed, or just to turn them off when they aren't wanted.

This also adds a second key binding to just turn off the mesh swaps
during the shadow pass, since the performance monitor shows this as the
most expensive. This can also disable the shadows, since it looks a bit
odd to have the costumes replaced but not the shadows.