DOAXVV/Help Shader: Add notification mechanism
This notification mechanism works by assigning a resource to
ResourceNotification, calling CustomShaderFormatText and setting z5 to
the time the notification is to disappear (e.g. z5 = time + 2).

These can easily be set from key bindings, as in the DOAXVV example
where it is set to two different messages when toggling costume mods on
and off, and three different messages when cycling shadow modes.

The notifications are only in the help shader as a temporary measure to
share the window size ini params - later, once we have ini param
independent variables and the ability to update custom resources they
will be split out.

This also fixes an issue where the final line of text would be
misaligned if it did not end in a newline, when aligning to center or
right.