LiSBtS / Auto-convergence: Disable remembering last frame convergence
In DOAXVV I found having the auto-convergence shader remember the
convergence value it had set in the last frame worked pretty well, but
in LiSBtS it ends up preventing the auto-convergence from working at
all.

This will probably need a bit more smarts in 3DMigoto and the ini file
to really resolve (e.g. updating the convergence value when it is
staged), so for now just disable this code.