Switch debug shaders to use BuiltInCommandListUnbindAllRenderTargets
This removes the dependency on a old d3dx.ini stanza, but depends on
newer versions of 3DMigoto that have this command list built in. These
versions have been out long enough to justify this, and if these are
used in older versions it's an easy enough change to revert.