Debug Shaders (2D + Cube): Switch to included ini
This simplifies importing these debug shaders into another project so
that they just have to be dropped in place and an include stanza added
for the desired shaders.

The downside of this is that instead of:

  ResourceDebug2D = cs-t0

It is now:

  Resource\ShaderFixes\debug_2d.ini\Debug2D = cs-t0

We might consider adding some form of section aliasing at a later data,
but for now since this is just for debugging it will do.