Lichdom: Better performing fix for Occlusion culling
This removes the deny_cpu_read hack, which had the effect of disabling
occlusion culling and fixing the flicker, but at the cost of losing the
performance benefits of using occlusion culling in the first place -
this was quite significant in one of the later levels in this game.

This replaces it with a custom shader that combines depth buffers from
the left and right eyes into a single mono depth buffer, allowing us to
keep the performance benefit of occlusion culling without (or with only
minor) flickering. This is currently performed part-way through the
game's downscaling process as a compromise for the performance of the
stereo2mono copy on SLI - refer to the notes in the ini for details.