3DMigoto Blender Addon: Add support for UNORM/SNORM formats
This adds support for importing and exporting signed and unsigned
normalised integer fields found in the vertex buffers of some games.
I've spotted these come up a few times in a random selection of vertex
buffers from various games, though this is presently adding support at
the request of a modder wanting to use the plugin to import vertex
buffers from the PS Vita version of DOAX3.