asmtool: Correct halo fix using HLSL in rare edge case
Ori and the Blind Forest has shaders that write to the Z component of
the output position separately to the rest, which triggers a rare code
path of the auto vertex halo fix that was not correctly ported to output
assembly when run from asmtool instead of hlsltool, resulting in a HLSL
assignment being inserted in the middle of an assembly shader.

This adds the abstraction to handle it in either HLSL or assembly,
including producing the correct swizzle when multiple components are
used in assembly.

Reported-by: Masterotaku