unity_asset_extractor: Fix resource table alignment in v14 + v15 files
DHR his this assertion in ReCore using the v15 file format, so apply the
same fix as for the v17 file format to align the file to a multiple of 4
bytes rather than assume a fixed three byte padding.

Also apply the same fix to the v14 parser as I am 99% certain it will be
the case there as well.