DOAXVV: Update material shader hashes for 2018-12-06 update
Yesterday's update added an extra parameter to a bunch of shaders,
changing their hashes and breaking the mods.

Most of the shaders don't actually use the new parameter, so calculating
the bytecode hash provided a way to map the old hash to the new hash. In
some ways it would be advantageous to switch to using the bytecode hash,
but doing so would invalidate any hashes that the update did not.

For those shaders that did change, using the new clipboard marking mode
was useful to rapidly collect the new hashes, and diffing the shaders
revealed a small amount of additional code near the end of the shader.

This also adds a bunch more clothing shaders added from various updates
while I'm at it. The fingernails shader is now also used for part of the
Chinese dress costume, so upgrade that to the full costume command list.