DOAXVV: Switch to using named variables, move HUD toggle to d3dx.ini
This reduces our use of IniParams for where named variables would
suffice - $costume_mods, $shadow_mode and $hud_visible are now all named
variables, the first two of which are persistent.

To avoid the need for an IniParam for hud_visible, move the HUD toggle
into the d3dx.ini command lists.