DOAXVV: Move auto-convergence settings into own shader
This moves the auto-convergence settings out of the d3dx.ini and into
the auto_convergence.ini as global variables. The IniParams are now only
set while auto-convergence is actually running so that modders don't
risk inadvertently breaking auto-convergence if they happen to pick a
variable it uses.

The auto-convergence shader still needs an update to move some control
logic out into the command lists, but that is a lower priority.