Help text shader: Remove IniParams reservation
This removes the reserved use of x5, y5 and z5 in the help text shader,
filling them in on demand and restoring the old values after freeing
them up for the shaderhacker to use for other purposes.

This also removes the unused SV_InstanceID input from the pixel shader
that was throwing a linkage error in the debug layer.