DOAXVV: Add frame analysis profiles
This adds a key binding Ctrl+F8 to cycle between three frame analysis
profiles:
- Dump skin and clothes
- Dump Skin, Clothes, Hair & Fingernails
- Dump UI

This just uses the fact that the analyse_options and dump commands are
in the command list, and can therefore be subject to conditional logic.
It sets variables for each case and runs the notification shader to
indicate which profile is active. It doesn't actually start frame
analysis, which is done by pressing F8 as before.