DOAXVV/Auto-Convergence: Move control logic into d3dx.ini, save popout bias
This is a fairly major update to the auto-convergence shader to move
much of the control logic out of the shader and into the d3dx.ini and
allow the enabled state and popout bias to be persistently saved to the
d3dx_user.ini file.

Some of the changes to support this may result in some observable
changes to the shader's behaviour - in particular, the popout bias
adjustment is calculated on the CPU instead of the GPU so that it can be
saved persistently, but this cannot calculate the equivalent popout bias
adjustment at this point so it just does a much simpler approximation.

There may also be some minor judder observable while adjusting the
popout bias, since the user convergence will be clobbered a frame or two
later once the auto-convergence shader has calculated a new target
convergence and staged it back to the CPU. This could be addressed by a
future 3DMigoto update to keep StereoParams in sync.

This also adds the actual convergence to the auto-convergence HUD, so
the user can see what it is doing, and allows the HUD to be kept on
screen via a setting.