extract_unity53_shaders: Add mapping for another "version 310 es" shaders
Also seen in Memories Retold. Not sure why both type 2 and 3 seem to
map to the same version, but don't particularly care either.

We're still missing the mapping for type 7 (and ofc any new ones added
after metal), which is presumably going to be another GL type based on
those around it, but leave it out until we actually find it in the real
world.

Reported-by: Losti