Add inverse shader variant with flexible offsets
In games where we need to do a lot of matrix inverses at different
offsets it's a bit of a pain to duplicate the inverse shader for each
one. This variant is designed to be more flexible, passing a parameter
in to tell it what offset to find the matrix in the constant buffer.