DOA6: Fix randomly detached body parts
The character models in this game are divided over a number of
sub-meshes, each drawn by different shaders. The driver can't seem to
make up its mind whether it should stereoise them or not, leading to
things like some characters in menus being stereoised and others not,
and random body parts detaching themselves from the character.

This heuristic seems to be decided at the time the shaders are created,
as once a body part is detached it will stay that way until it is
unloaded, but the heuristic seems quite inconsistent, as can be seen by
flipping through Nyotengu's costume options in the menu a few times.

This adds a ShaderRegex that matches all character models (skin,
clothes, etc), disables the driver's stereo correction, and implements
our own to replace it.