DOA6: Replace driver stereo correction with our own everywhere
The driver is just totally confused by this game, and while it might be
possible to find some combination of StereoTextureEnable or
StereoFlagsDX10 that work with it, let's scrap the driver's stereo
correction entirely and replace it with our own, including heuristics.

This removes the match on bSkin so that the character regex will match
every single vertex shader. So as not to totally break the game we add
two heuristics of our own - a depth buffer must be bound and we do not
stereo correct when the output depth is exactly 1.0.

With this change all 2D/3D mis-matched geometry caused by shaders seems
to have been resolved, though there are some halo issues and the like to
deal with.