DOA6: Force right side match up screen render targets to stereo
This is another case of inconsistency with this game - the render
targets for the character on the left side of the match up screen are
always stereo, but the render targets on the right side can't seem to
make up their minds - sometimes they are stereo, sometimes mono, and
sometimes some broken hybrid of the two.

The right hand side render targets are square, which will be hitting the
square render target heuristic in the driver, hence why it is making
them mono. Add some fuzzy matches looking for square render targets with
the colour formats used by these render targets to force them to stereo.
Alternatively we should be able to force the render targets on the left
to mono, but as they are rectangular targets they have a greater risk of
colliding with something unrelated.