DOA6: Fix Halos via Regex
This splits the StereoCorrection regex in two, with one case to handle
simple halo patterns, and the second to handle other vertex shaders.

The halo variant works by matching shaders that use a mov instruction to
copy a temporary register to the output position (indicating that it
will be used somewhere else as well), and applying the stereo correction
before the mov. It won't handle as many variations on this as asmtool
can, but it works well enough for this game.

The remaining variant matches all shaders that used something other than
a mov instruction to write to the output position, as this suggests that
the output position was not used anywhere else. This swaps the output
position register for a temporary and adjusts it before the ret at the
end of the shader.

In order to prevent the halo variant matching a shader already modified
by the non-halo variant (since that will switch the output to a mov,
which the halo variant matches) it inserts a comment in the SV_Position
declaration to stop the other regex matching.