DOA6: Convert point light fix into ShaderRegex
This moves the point light fix into ShaderRegex to fix other lights. The
biggest challenge here is that the offset of the mW2P and mP2W matrices
is different in each shader, so this matches the offsets, does an
integer division by 16 in the shader and uses it as a dynamic offset to
the $Globals constant buffer.

Directional shadows are left as a ShaderFix for now since they lack one
of the matrices the regex will be a little more involved.