DOA6: Fix clouds (unusual pattern to avoid clipping, stretching & smearing)
This game uses ray-traced clouds, and while the usual skybox approach
does move them to infinity, it also introduces clipping on the sides of
the screen. This can be avoided by stretching the sky, but that's still
not ideal.

The vertex shader is passed four camera vectors for the corners of the
screen that will be interpolated by the rasterizer to pass the pixel
shader a vector going through it's pixel. This adjusts these vectors to
take into account the off-center stereo projection, which results in the
clouds being moved to infinity without clipping or stretching. We can
use the fact that the vectors corresponds to the corners of the screen
to convert separation into their scale and orientation, without having
to know what that is.

This also corrects the delta position buffer output from the pixel
shader to eliminate smearing from previous frames. This is just a
bog standard stereo correction in clip space.