DOA6: Fix sun & flares
This uses a regex pattern to adjust the g_vSunScreenPos and
g_vMoonScreenPos to move the sun to infinity. It also adjusts the flare
shader to move the coordinates the flares check for to see if the sun is
obscured, and moves the flare coordinates.

There is a commented out ShaderRegex that can partially adjust the
flares, but it seems like the flare shaders interact with each other so
for now doing a simpler fix for flares in a single shader.