DOA6: Fix tiled decal clipping
First tile lighting, now tile decals.

Similar to tile lighting, a compute shader is used to decide which tiles
any given decal may be drawn in, and we need to modify those bounds to
allow it to be drawn in stereo.

This is a rough fix that works by adding separation to the min/max tile
positions so that the decals will be drawn so long as they are inside
the screen (they may still clip if popped out too far).