DOA6: Fix car windscreen
This car is found in one of the hidden areas in the DOA6 casino
downstairs in the museum stage (the one that starts in the warzone)

This shader is used with no bound pixel shader (depth pass I assume),
causing the driver heuristics to only run it for the right eye. Bind a
dummy render target to render it in both.