DOA6: Extend lighting pattern to fix mW2P usage
The changes to the lighting shader in the arena added a new broken
shadow on the floor that is fixed by correcting the uses of mW2P later
in the shader. Extend the regex to do this, though because mW2P's
location is highly variable in these shaders just match on code patterns
and hope for the best.