DOA: Force in-game 3D TVs to mono
The arena stage has several large displays showing the footage of the
game currently being played, and by default they themselves render in
3D, so you end up with a 3D TV inside the same 3D TV inside the same 3D
TV inside your monitor.

Neat - but it tends to violate infinity since the stereo projection is
calibrated for the top level monitor, not any of the in-game monitors.

Force the image on these monitors to mono to make them more comfortable
to view, but leave a setting under [Constants] in case a user would like
to make them stereo again.