DOA6: Fix snow shadows / decals
These are the shadows underneath cars in the snowing stage, and the
pattern is nearly identical to the decal pattern, so implement it as a
regex. Like the decals, these shadows clip on tile boundaries, so also
adjust the compute shader responsible for them to extend them by
separation to remove clipping (provided they aren't popped out too far).