DOA6: Fix halo in river water
This requires a world-space correction be applied to one of the other
outputs from the vertex shader.

This shader only has the mW2P matrix, but we also need a mP2W or similar
to change the world space coordinate. The directional shadow shader is
already using the flexible inverse shader to go the other way, and we
don't want to disable the per frame limits since this shader is called
numerous times. Modify the flexible inverse shader to add several slots,
each of which will have independent per-frame limits, thereby allowing
it to be used to inverse each unique matrix once.

This also adds the flexible shader to the dedicated directory that was
missed in commit 723c300a09ad109d3bb8888d93816c239b21faf4