DOA6: Fix environment map reflections on water
This shader controls both reflection and refraction on the river. It
needs the usual kind of correction for reflections (stereo correct after
sampled depth to move to screen depth, followed by an eye position
correction to move them to accurate depth), but doing so breaks
refraction, so do not adjust the original r0 register, but do adjust the
copy that takes the eye position into account.