DOA6: Fix physics desync
The physics can desync while 3D Vision is disabled because it doesn't
run in the left eye while 3D Vision is disabled, and will be out of sync
with the right eye when 3D Vision is re-enabled (and never resets unless
restarting the whole game). This is particularly evident in the lost
paradise stage as all the plants desync.

The fix is to perform a mono copy of the previous physics phase buffer.
Forcing the buffer to mono should also work, but risks breaking other
things (forcing all buffer UAVs breaks the A.P.O. level), and while I
may be able to dial it in, it's easier to just do the mono copy fix for
now.