DOA6: Fix coins & random graphical glitches in pirate ship
The coin particles are controlled by compute shaders that populate the
vertex buffers and draw call parameters to a later graphics draw call.
The compute shader's UAVs get out of sync between each eye, leading to
inconsistent particles between the eyes as well as random white flashes.

I'm not entirely sure how the buffers get out of sync in the first place
since all the inputs seem to match up between the eyes, but there might
be some difference, or it could be similar to the wind physics buffer
that persists between frames and the left eye doesn't get updated with
3D Vision off...

To solve this, we step in after the last compute shader has done it's
work, create mono copies of the vertex buffer & parameter UAVs, then use
a custom shader to copy them back to the stereo UAVs that the game holds
and will later pass to the graphics pipeline.