DOA6: Fix specular reflections on main menu / character select
The shaders used on the menu have a different pattern to the rest of the
lighting shaders and the specular is applied during the forward pass.
These ones are a little tricky, because they don't have the matrices
needed to make the adjustment nor anything else we can use. The vertex
shaders do, so add a regex to match the vertex shaders and grab their
matrix if it's in the expected position, then inverse it and use it in
the pixel shader.

We don't do the inverse in the vertex shader command list because that
regex will match a lot of extra shaders where the matrix doesn't match.
Doing it from these will break the rendering if the inverse limit is in
place, and removing the inverse limit harms performance way too much.
Doing the inverse in the pixel shader allows us to do it only when
needed, and use the limit to avoid any performance hit.