DOA6: Fix green water pool in hidden garden stage
This is a slight variant of the same pattern as the river from the
dinosaur stage, though the halo is only fixed in the right eye - likely
indicating the presence of a mono render/depth target somewhere.

One difference to the river shader was that the depth texture it samples
from was mono, requiring the shaders that originally wrote to it to be
forced stereo (via dummy render target assignment) and textureoverrides
to force the depth target and texture to stereo.