DOA6: Move 3D Vision ini changes into a mod subdirectory
If this ends up being used like DOAXVV some people might want to disable
the 3D Vision components since they do add some overhead. Move them all
into a "Mods/3d-vision" directory to make it easy to disable them all at
once.

Add checks to all shaders in ShaderFixes that use an injected cb13, so
that they can render in 2D without it since that will be missing if the
3D Vision mod is disabled.

Add the 3d-vision mod to the uninstall.bat. We don't usually want to
remove a user's mods since they are installed separately, but this is
one that will (maybe) be installed with the base mod, so it counts.