DOA6: Disable menu adjustment in story mode
I haven't finished story mode as yet, and there are new backgrounds
every three or so matches. The 3D character portrait isn't displayed in
story mode, so it's not necessary to adjust it at all. Change the title
adjustment to only run when in a menu with a known background to avoid
half-adjusting story, and remove all story mode background hashes.

Commit some missing entries from d3dx.ini