DOA6: Change screenshot button to XB_BACK
Pressing the thumbstick inwards resets the camera in the victory screen,
so no good using that for a screenshot button. Guide is my next usual
choice, but can be difficult to ship and explain it to people thanks to
the Windows 10 game bar interfering. The game doesn't seem to use the
back button on the controller for anything, so let's try that.