DOAXVV: Auto-adjust pool hop minigame icon depth
This adds an automatic depth adjustment to the icons in the pool hopping
minigame. As usual the game doesn't give us the depth of HUD icons, so
this includes a variant of crosshair.hlsl adjusted for this game to ray
trace the depth buffer and locate a suitable depth.

The icons are made up of several pieces (bubble, item icon, quantity
text), and in order to adjust them all to the same depth we match the
bubble texture and capture the matrices used with it. We then calculate
the screen coordinate based on passing local coordinates (0,0) through
those matrices, which gives us the anchor point on the pool platform
below the icon, which is a perfect position to ray-trace on.