Inverse shader: Copy input matrix to output & preserve slot results
For convenience this copies the input matrix to the output, so that the
pair of forward and inverse matrices can be accessed together.

It also removes the common CB copy target, and makes each slot's CB
resource a copy target instead. This allows the common path to be used
if a copy is not wanted (e.g. if accessing the resulting matrix via a
shader resource or UAV slot).