DOA6: Improve coin fix
The buffers I identified earlier were not the only buffers used for the
coin effect. This switches the hash based match to a fuzzy match based
on matching the drawindirect_args misc flag, or the vertex_buffer bind
flag. It also includes a commented out method that can be used to match
all indirect draw calls in 3DMigoto 1.3.16 and force buffers to mono at
the last minute.