3DMigoto Blender Addon: Continue with next mesh on 4D positions/normals
These assertions are triggering on some of the meshes in DOA6, stopping
the import prematurely. Disabling the assertions shows that the meshes
triggering them are oriented incorrectly (oriented 90 degrees from where
they should be - axis swap?), so it doesn't seem safe to disable the
assertion, but we still want the rest of the meshes to import.

Turn the assertion into a Fatal exception, which is already caught and
will continue to the next mesh. Add the filename to the error message so
we can see which meshes are triggering this.