3DMigoto Blender Addon: Allow 1/2/4 channel COLORs
This allows COLOR to be used with any number of fields, which is useful
to import and export arbitrary / unknown data and perhaps visualise it
over the mesh. 1 and 2 channel COLOR elements simply have the remaining
channels omitted, while 4 channel COLORs are split between RGB and A
channels allowing the alpha to be preserved even though Blender doesn't
support vertex alpha.