3DMigoto Blender Addon: Import int layers out of signed range
DOA6 has some PSIZE/FOG values set to 4294967295, as in UINT32_MAX.
However Blender's integer vertex layers are 32bit signed values
that cannot store that value and throw an exception when it is assigned.
This solves the issue by reinterpreting large integers as signed values
so that they will import. Export works as is with no changes requires.