Debug Shaders: Use namespace & shorted Resource names
The debug shaders are a bit verbose to use, so this overrides their
namespace to be shorter, and removes the redundant "Debug" from the
resource names. So, rather than:

    Resource\ShaderFixes\debug_cb.ini\DebugCB = ps-cb0

You would now do:

    Resource\debug_cb\CB = ps-cb0

This also removes the absolute path names to the HLSL files since
3DMigoto can find them next to the ini files, and applies these changes
to all uses in DOA6.