DOA6: Add script to decode soft body nodes from g1m files
Editing certain meshes in DOA6 will need the soft body weights
recalculated, which will require knowledge of the positions of all the
soft body nodes. These can be obtained from Frame Analysis, but the
positions will have the physics sim already applied and be in world
space, and we really want the original positions in local space.

This script parses the SOFT section of the g1m files and will hopefully
provide the position data we need. It is not yet integrated into
Blender, and currently the SOFT chunk must be extracted with a hex
editor, but it is able to parse the data structure in at least one mesh.