decode_doa6_soft: Decode g1m files directly
This adds a simple G1M parser to locate the SOFT section from these
files rather than having to extract it manually with a hex editor.

Removes an assertion on one of the header fields that varies in other
files and we already determined indicates the length of one of the lists
in the file.