decode_doa6_soft: Add Blender import operator
This allows the script to be used as a Blender addon to import the soft
body nodes into a scene, that we will later be able to use to update the
soft body mesh PSIZE, FOG and related TEXCOORDs to allow vertex
positions to be properly altered.

The soft body nodes are imported as a series of cubes to keep things
simple, and they have an empty parent used to group nodes in the same
region.

The script will still work standalone - if bpy cannot be imported it
will skip all Blender integration and just print out the decoded format
instead.