decode_doa6_soft: Add operator to update vertex soft body weights
This adds an operator that will update the PSIZE, FOG, TEXCOORD8 and
TEXCOORD9 components of every vertex in the selected mesh(es) based on
their relative positions to the surrounding nodes in the soft body grid
that must also be selected.

For each vertex, up to 8 surrounding nodes are located from the soft
body grid (ensuring that they are all in different directions). The
vertex position is then compared to these to calculate the amount each
node position will contribute to the vertex position and the attributes
are updated.

Currently this weight calculation is just an approximation, and the
resulting mesh positions are noticeably incorrect - the calculation will
have to be fine tuned to make this usable, but it is close enough to
prove this method does work, and this does allow the vertex positions to
be altered.