decode_doa6_soft: Implement trilinear interpolation for accurate results
This replaces the fudged proximity based interpolation with more
accurate trilinear interpolation, which gives significantly more
accurate results and a smooth mesh practically indistinguishable from
the original but now with the vertex positions altered.

To achieve this it takes the 8 surrounding nodes and locates 4 parallel
sides, then it interpolates along them using a ratio based on the
closest point to those lines. This gives a square that it interpolates
along two parallel sides giving a line, that it interpolates along one
last time to find a point.

In addition to interpolating the locations of the nodes, it assigns
weights at each node (like [1,0,0,0,0,0,0,0]) that is interpolated using
the same ratios as for the position, and the resulting interpolated
weights are then assigned to the vertex, which is far more accurate than
the distance method, and doesn't require a fudge factor.