3DMigoto Blender Addon: Allow DOA6 poses to be imported (buggy)
DOA6 stores their bone matrices in a constant buffer shared with other
data, so to allow them to be imported this adds a new property to the
import dialog to specify the start and end offsets within the constant
buffer.

This only sort of works - the bones are imported and their matrices
match the file, but the mesh is heavily distorted and I have yet to work
out exactly what the issue is. Despite this, it should be sufficient to
merge the meshes together and match up the vertex groups, and examining
the weights after performing a merge operation does suggest that it is
working.

It also fixes up a few issues in the merge operator affecting DOA6 where
the some of the meshes seem to be using vertex groups with no
corresponding bone, triggering a KeyError.