Hack Blender Addon to remove Semantic processing for DOA6 skirts
The skirt meshes in DOA6 cannot import successfully into Blender for
numerous reasons - the positions and normals are 4 dimensional, they
have a BINORMAL semantic the script doesn't yet support, their TEXCOORD5
is an integer semantic, their COLOR1 has negative values, their
BLENDWEIGHT semantic has negative values, and their NORMAL, TANGENT and
BINORMAL also look suspect. Pretty much the only semantic that actually
looks sane is TEXCOORD0.

In order to import and export these to Blender, almost all the usual
semantic processing has to be disabled and just about everything has to
be imported and exported via vertex layers (including the W component of
the vertex positions). Even once they are imported the meshes are quite
difficult to edit, because the 4D coordinates are wrapped in on each
other - to the point where it is easier to work with the UV layer (the
one that actually contains UV data) rather than the 3D projection of the
4D mesh.

This commit seriously nerfs the Blender addon to achieve this, and will
remain in a topic branch for the time being.

Going forward the script could be made to detect some of these issues
and fall back to using vertex layers for any semantic with data that
does not match the expected ranges, but even this probably still won't
cover NORMAL or TANGENT, which look valid - but damage the mesh if
imported and exported with the usual processing.