Blender addon: Save all unhandled semantics in vertex layers
This removes the explicit list of semantics to save in vertex layers and
default to saving any unknown semantic in a vertex layer. This will
hopefully allow a wider variety of meshes to be successfully imported &
exported even if they use exotic semantics (provided that they are
telling the truth about any of the semantics that the script has
explicit support for, which is not the case for some DOA6 skirt meshes).

This also adds a check that TEXCOORD semantics do contain the expected
floating point data, and if not they will be stored in vertex layers
instead as that would indicate that they contain arbitrary data rather
than UV layers.

This also adds support for storing the W coordinate of a 4D POSITION
semantic in a vertex layer, however this is disabled as meshes using 4D
coordinates (such as the skirts in DOA6) may have other more severe
problems and preserving this by itself may not be enough to import &
export these (it certainly is not for the DOA6 skirts). The fatal error
can be commented out to try preserving the W coordinate to try this, or
the version of the script found in the blender_dumbed_down may be used
as it has disabled almost all semantic processing and therefore is less
likely to damage meshes that lie about their semantics, at the cost of
not being able to edit things like normals.