3DMigoto Blender Addon: Add vertex group step for DOA6 poses
The vertex groups in DOA6 are pre-multiplied by three to match up with
their offsets within the bone matrices. This adds a property to set the
vertex group step when importing a pose, which allows the DOA6 poses to
be imported and the merge poses operator now appears to be working
correctly on these.