DOA6: Add bone debug visualisation shader
This adds a visualisation debug shader to show the transformation of the
bones in a given draw call, and can be used with something like this:

    [TextureOverrideBody]
    hash = 797e7cc8
    if ps !== -0.0
    	$\debug_bones\first = 0
    	$\debug_bones\last = 72
    	run = CustomShader\debug_bones\VisualiseBones
    endif

    [TextureOverrideBoots]
    hash = 0cafc908
    if ps !== -0.0
    	$\debug_bones\first = 0
    	$\debug_bones\last = 27
    	run = CustomShader\debug_bones\VisualiseBones
    endif

Note that the visualisation is not intuitive - the dots representing the
bones are not positioned where you would naively expect. They actually
show how far from the origin any region of mesh would be moved, but they
are relative to the origin, while the mesh is relative to the mesh.